feat: add O(n+m) search in row-column sorted matrix
